Refactor request handling and update naming conventions
Renamed `RouteEndpointBuilderExtensions` to `EndpointRouteBuilderExtensions` and added `MapGetQuery` method for handling GET requests with query parameters. Removed `using Infinity.Toolkit.Experimental;` in `HandlerContext.cs` and restored the definitions of `IRequestHandler<TIn>` and `IRequestHandler<TIn, TResult>`.

Updated `ServiceCollectionExtensions.cs` to replace `AddMediator` with `AddRequestHandler`, adjusting return types and logic accordingly. Renamed `MediatorHandlerBuilder` to `RequestHandlerBuilder` and updated related extension methods to reflect the new naming.

These changes shift the codebase from a mediator pattern to a more direct request handler approach, enhancing clarity and maintainability.
