+++ src/Infinity.Toolkit.Handlers/ServiceCollectionExtensions.cs
@@ -1,32 +1,22 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace Infinity.Toolkit.Handlers;
 
 public static class ServiceCollectionExtensions
 {
     /// <summary>
-    /// Adds the mediator to the service collection.
-    /// </summary>
-    public static IServiceCollection AddMediator(this IServiceCollection services)
-    {
-        services.TryAddTransient<IMediator, Mediator>();
-        return services;
-    }
-
-    /// <summary>
-    /// Adds a mediator handler to the service collection.
+    /// Adds a request handler to the service collection.
     /// </summary>
     /// <typeparam name="TRequest">The request type.</typeparam>
-    /// <typeparam name="TMediatorHandler">The mediator handler type.</typeparam>
+    /// <typeparam name="TRequestHandler">The mediator handler type.</typeparam>
     /// <param name="services">The service collection.</param>
-    /// <returns>A <see cref="MediatorHandlerBuilder"/> instance used to configure the mediator handler.</returns>
-    public static MediatorHandlerBuilder AddMediatorHandler<TRequest, TMediatorHandler>(this IServiceCollection services)
-        where TRequest : class, ICommand
-        where TMediatorHandler : class, IRequestHandler<TRequest>
+    /// <returns>A <see cref="RequestHandlerBuilder"/> instance used to configure the request handler.</returns>
+    public static IServiceCollection AddRequestHandler<TRequest, TRequestHandler>(this IServiceCollection services)
+        where TRequest : class
+        where TRequestHandler : class, IRequestHandler<TRequest>
     {
-        services.AddTransient<IRequestHandler<TRequest>, TMediatorHandler>();
-        return new MediatorHandlerBuilder(services, typeof(TRequest));
+        services.AddTransient<IRequestHandler<TRequest>, TRequestHandler>();
+        return services;
     }
 
     /// <summary>
@@ -34,15 +24,16 @@ public static class ServiceCollectionExtensions
     /// </summary>
     /// <typeparam name="TRequest">The request type.</typeparam>
     /// <typeparam name="TResult">The result type.</typeparam>
-    /// <typeparam name="TMediatorHandler">The mediator handler type.</typeparam>
+    /// <typeparam name="TRequestHandler">The mediator handler type.</typeparam>
     /// <param name="services">The service collection.</param>
-    /// <returns>A <see cref="MediatorHandlerBuilder"/> instance used to configure the mediator handler.</returns>
-    public static MediatorHandlerBuilder AddMediatorHandler<TRequest, TResult, TMediatorHandler>(this IServiceCollection services)
-        where TRequest : class, IQuery
-        where TMediatorHandler : class, IRequestHandler<TRequest, TResult>
+    /// <returns>A <see cref="RequestHandlerBuilder"/> instance used to configure the mediator handler.</returns>
+    public static IServiceCollection AddRequestHandler<TRequest, TResult, TRequestHandler>(this IServiceCollection services)
+        where TRequest : class
+        where TResult : class
+        where TRequestHandler : class, IRequestHandler<TRequest, TResult>
     {
-        services.AddTransient<IRequestHandler<TRequest, TResult>, TMediatorHandler>();
-        return new MediatorHandlerBuilder(services, typeof(IRequestHandler<TRequest, TResult>));
+        services.AddTransient<IRequestHandler<TRequest, TResult>, TRequestHandler>();
+        return services;
     }
 
     /// <summary>
@@ -50,9 +41,13 @@ public static class ServiceCollectionExtensions
     /// </summary>
     /// <typeparam name="TRequest">The request type.</typeparam>
     /// <typeparam name="TDecorator">The decorator type.</typeparam>
-    public static IServiceCollection DecorateMediatorHandler<TRequest, TDecorator>(this IServiceCollection services)
+    public static IServiceCollection DecorateRequestHandler<TRequest, TDecorator>(this IServiceCollection services)
+        where TRequest : class
         where TDecorator : class, IRequestHandler<TRequest>
-        where TRequest : class, ICommand => services.Decorate<IRequestHandler<TRequest>, TDecorator>();
+    {
+        services.Decorate<IRequestHandler<TRequest>, TDecorator>();
+        return services;
+    }
 
     /// <summary>
     /// Decorates all registered instance of <see cref="IRequestHandler{TQuery, TQueryResult}"/> using the specified type <typeparamref name="TDecorator"/>.
@@ -61,39 +56,44 @@ public static class ServiceCollectionExtensions
     /// <typeparam name="TRequest">The request type.</typeparam>
     /// <typeparam name="TResult">The result type.</typeparam>
     /// <typeparam name="TDecorator">The decorator type.</typeparam>
-    public static IServiceCollection DecorateMediatorHandler<TRequest, TResult, TDecorator>(this IServiceCollection services)
+    public static IServiceCollection DecorateRequestHandler<TRequest, TResult, TDecorator>(this IServiceCollection services)
+        where TRequest : class
+        where TResult : class
         where TDecorator : class, IRequestHandler<TRequest, TResult>
-        where TRequest : class, IQuery => services.Decorate<IRequestHandler<TRequest, TResult>, TDecorator>();
+    {
+        services.Decorate<IRequestHandler<TRequest, TResult>, TDecorator>();
+        return services;
+    }
 }
 
 /// <summary>
-/// A builder used to configure a mediator handler.
+/// A builder used to configure a request handler.
 /// </summary>
 /// <param name="services"></param>
 /// <param name="handlerType"></param>
-public sealed class MediatorHandlerBuilder(IServiceCollection services, Type handlerType)
+public sealed class RequestHandlerBuilder(IServiceCollection services, Type handlerType)
 {
     internal IServiceCollection Services { get; } = services;
 
     internal Type HandlerType { get; } = handlerType;
 }
 
-public static class MediatorHandlerBuilderExtensions
+public static class RequestHandlerBuilderExtensions
 {
     /// <summary>
     /// Decorates all registered instances of <see cref="IRequestHandler{TRequest}"/> using the specified type <typeparamref name="TDecorator"/>.
     /// </summary>
     /// <typeparam name="TDecorator">The decorator type.</typeparam>
-    /// <returns>A <see cref="MediatorHandlerBuilder"/> instance used to configure the mediator handler.</returns>
-    public static MediatorHandlerBuilder Decorate<TDecorator>(this MediatorHandlerBuilder builder)
+    /// <returns>A <see cref="RequestHandlerBuilder"/> instance used to configure the mediator handler.</returns>
+    public static RequestHandlerBuilder Decorate<TDecorator>(this RequestHandlerBuilder builder)
         where TDecorator : class => builder.Decorate(typeof(TDecorator));
 
     /// <summary>
     /// Decorates all registered instances of <see cref="IRequestHandler{TRequest}"/> using the specified type <paramref name="decoratorType"/>.
     /// </summary>
     /// <param name="decoratorType">The decorator type.</param>
-    /// <returns>A <see cref="MediatorHandlerBuilder"/> instance used to configure the mediator handler.</returns>
-    public static MediatorHandlerBuilder Decorate(this MediatorHandlerBuilder builder, Type decoratorType)
+    /// <returns>A <see cref="RequestHandlerBuilder"/> instance used to configure the mediator handler.</returns>
+    public static RequestHandlerBuilder Decorate(this RequestHandlerBuilder builder, Type decoratorType)
     {
         builder.Services.Decorate(builder.HandlerType, decoratorType);
         return builder;
